proto: move IBC-rs impls into foreign impls section
