+++ proto/src/lib.rs
@@ -56,23 +56,6 @@ pub mod client {
 /// IBC protocol structures.
 pub mod ibc {
     tonic::include_proto!("penumbra.ibc");
-
-    extern crate ibc as ibc_rs;
-
-    use ibc_rs::core::ics02_client::client_state::AnyClientState;
-
-    use crate::Protobuf;
-    use ibc_proto::google::protobuf::Any;
-    use ibc_proto::ibc::core::channel::v1::Channel as RawChannel;
-    use ibc_proto::ibc::core::connection::v1::ConnectionEnd as RawConnectionEnd;
-    use ibc_rs::core::ics02_client::client_consensus::AnyConsensusState;
-    use ibc_rs::core::ics03_connection::connection::ConnectionEnd;
-    use ibc_rs::core::ics04_channel::channel::ChannelEnd;
-
-    impl Protobuf<RawConnectionEnd> for ConnectionEnd {}
-    impl Protobuf<Any> for AnyClientState {}
-    impl Protobuf<Any> for AnyConsensusState {}
-    impl Protobuf<RawChannel> for ChannelEnd {}
 }
 
 /// Wallet protocol structures.
